pyproject.toml
[project]
name = "siren-codec"
version = "0.1.0"
description = "SIREN: Semantic Information Representation for Efficient Neural-coding (~68 bps speech codec)"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "transformers>=4.30.0",
    "numpy<2.0",
    "scipy>=1.10.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "soundfile>=0.12.0",
    "einops>=0.6.0",
    "protobuf>=4.0.0",
    "matplotlib>=3.7.0",
    "datasets>=4.5.0",
    "speechmos",
    "librosa",
    "onnxruntime",
    "torchcodec",
    "tensorboard",
    "vocos>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ultra_low_bitrate_codec"]

[project.scripts]
train = "ultra_low_bitrate_codec.train:main"
download-data = "ultra_low_bitrate_codec.download_data:main"
preprocess = "ultra_low_bitrate_codec.preprocess_data:main"
train-bitnet = "ultra_low_bitrate_codec.train_bitnet_entry:main"
